pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ez_cqrs"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
authors = [
    { name = "Tomas Perez Alvarez", email = "tomasperezalvarez@gmail.com"}
]
dependencies = [
    "mashumaro",
    "result"
]

[project.optional-dependencies]
dev = [
    "pyrgo",
    "pytest-cov"
]

[tool.ruff]
line-length = 88
ignore = [
    "ANN101",
    "D203",
    "D212",
]
select = ["ALL"]
fix = false
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.per-file-ignores]
"scripts/*.py" = ["INP001"]
"__init__.py" = ["D104"]
"tests/*.py" = ["INP001", "S101"]

[tool.ruff.isort]
known-first-party = ["ez_cqrs"]
force-wrap-aliases = true
combine-as-imports = true
required-imports = ["from __future__ import annotations"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.flake8-quotes]
inline-quotes = "double"


[tool.black]
line-length = 88
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",
    "--strict-markers",
    "--cov-config=pyproject.toml",
    "--cov-fail-under=20",
    "--cov=ez_cqrs",
    "--cov-report=term-missing:skip-covered",
]
markers = [
    "integration: mark integration tests.",
    "unit: mark unittest.",
]

[tool.mypy]
show_error_codes = true
follow_imports = "normal"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true
